Creating ManageNotAcceptedRegisters style file
