+++ src/screens/manageCounselors/ManageNotAcceptedRegistersScreen.js
@@ -1,6 +1,5 @@
 import React from 'react';
 import {
-  StyleSheet,
   Text,
   View,
   ScrollView,
@@ -14,38 +13,7 @@ import LoadingIndicator from '../../components/LoadingIndicator';
 import Button from '../../components/Button';
 import { getManagerCounselorData } from '../../services/extractDataCounselor';
 import ScheduleCard from '../../components/ScheduleCard';
-
-const buttonBoxGreen = StyleSheet.create({
-  design: {
-    borderColor: 'black',
-    borderWidth: 1,
-    borderRadius: 7,
-    backgroundColor: '#4CD964',
-    padding: 8,
-    justifyContent: 'center',
-    marginRight: 13,
-  },
-  text: {
-    fontSize: 12,
-    textAlign: 'center',
-  },
-});
-
-const buttonBoxRed = StyleSheet.create({
-  design: {
-    borderColor: 'black',
-    borderWidth: 1,
-    borderRadius: 7,
-    backgroundColor: '#FF3B30',
-    padding: 8,
-    justifyContent: 'center',
-    marginRight: 13,
-  },
-  text: {
-    fontSize: 12,
-    textAlign: 'center',
-  },
-});
+import { buttonBoxGreen, buttonBoxRed} from '../../Styles/ManageConselors/ManageNotAcceptedRegistersStyles'
 
 export default class ManageNotAcceptedRegistersScreen extends React.Component {
   componentWillMount() {
